17-Exemplo/requirements.txt
flet-embed==0.23.1
